added ItemsMixin nodes
