+++ src/nodebpy/nodes/geometry/__init__.py
@@ -21,7 +21,6 @@ from .manual import (
     ClosureInput,
     ClosureOutput,
     ClosureZone,
-    FormatString,
     JoinStrings,
     Menu,
     Collection,
@@ -69,6 +68,7 @@ from .converter import (
     FilterList,
     FindInString,
     FloatToInteger,
+    FormatString,
     GetBundleItem,
     GetListItem,
     GetNestedBundlePaths,
